refactor: extract find_nearest_viewpoint and sample_navmesh to module level
Move these utilities from class methods / script-local functions to
module-level functions in habitat.py that accept a raw habitat.Env.
The wrapper class delegates to them, and the collection script /
test import them directly. Eliminates 3 duplicate implementations.
